backend/requirements.txt
fastapi
uvicorn
python-multipart
paho-mqtt
sqlmodel
requests
pydantic-settings
